Refactor files structure
