+++ pages/game/index.tsx
@@ -1,39 +1,29 @@
 import { NextPage } from 'next';
 import { getSession } from 'next-auth/react';
-import GameLayout from '../../components/game/game';
-import { getPlaylistGame } from '../../server/playlist';
-import { useDispatch, useSelector } from 'react-redux';
-import { useEffect } from 'react';
-import { gameSlicesActions } from '../../store/game-slice';
-import { LyricGame } from '../../interfaces/game';
 import Metadata from '../../components/meta/metadata';
-import { RootState } from '../../interfaces/state';
+import { getUserPlaylist } from '../../server/playlist';
+import { PlaylistSelection } from '../../interfaces/playlist';
+import PlaylistLayout from '../../components/game/playlists';
 
 interface Props {
-  playlist: LyricGame[];
+  playlists: PlaylistSelection[]
 }
 
-const Auth: NextPage<Props> = ({playlist}) => {
+const Playlist: NextPage<Props> = ({playlists}) => {
 
-  const dispatch = useDispatch();
-  const gameState = useSelector((state: RootState) => state.game);
-
-  useEffect(() => {
-    dispatch(gameSlicesActions.setGameState({playlist, playlistName: 'Liked songs', currentSong: 0}));
-  }, [dispatch, playlist]);
-  
   return (
     <>
-      <Metadata title={`Game | ${gameState.playlistName}`}/>
-      <GameLayout />
+      <Metadata title='Playlist | Playlist Master' description='Select the playlist to play' />
+      <PlaylistLayout playlists={playlists} />
     </>
   );
 };
 
 export const getServerSideProps = async(context: any) => {
   const session = await getSession({req: context.req});
-  let playlist = await getPlaylistGame(session?.accessToken, Number(process.env.LIMIT_SONGS) || 10);
-  if(!session || !playlist) {
+  let playlists = await getUserPlaylist(session?.accessToken, Number(process.env.LIMIT_PLAYLISTS) || 20);
+
+  if(!session || !playlists) {
     return {
       redirect: {
         destination: '/auth',
@@ -41,15 +31,15 @@ export const getServerSideProps = async(context: any) => {
       }
     }
   }
-  
-  playlist = playlist.sort(() => 0.5 - Math.random());
-  
+
+  playlists = playlists.sort(() => 0.5 - Math.random());
+
   return {
     props: { 
       session,
-      playlist
+      playlists
     }
   }
 };
 
-export default Auth;
+export default Playlist;
